requirements.txt
requests 
pandas 
python-dotenv
numpy 
scikit-learn
pydantic
fastapi
pymongo